requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
distlib==0.3.8
distro==1.9.0
Django==5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
easy==0.0.4
filelock==3.13.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-cloud-translate==3.15.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googletrans==4.0.0rc1
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
Jinja2==2.7.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.1
numpy==1.26.4
openai==1.23.3
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.39
Pillow==10.1.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==16.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==3.11
referencing==0.35.0
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
setuptools==69.0.3
sh==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.4.4
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.13
yarl==1.9.4
